Change access and move MessageSinkMappingImpl to its own file
